services/fastapi-userservice/pyproject.toml
[tool.poetry]
name = "user-service-fastapi"
version = "0.1.0"
description = ""
authors = ["subhamsahu <sam766626@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aioredis = "2.0.1"
aiosmtplib = "1.1.6"
anyio = "3.6.1"
async-timeout = "4.0.2"
autopep8 = "1.6.0"
bcrypt = "3.2.2"
blinker = "1.4"
certifi = "2022.6.15"
cffi = "1.15.1"
charset-normalizer = "2.1.0"
click = "8.1.3"
cryptography = "37.0.2"
Deprecated = "1.2.13"
dnspython = "2.2.1"
ecdsa = "0.17.0"
email-validator = "1.2.1"
fakeredis = "1.8.1"
fastapi = "0.78.0"
fastapi-mail = "1.0.9"
gunicorn = "20.1.0"
h11 = "0.12.0"
httpcore = "0.15.0"
httptools = "0.4.0"
httpx = "0.23.0"
idna = "3.3"
Jinja2 = "3.1.2"
MarkupSafe = "2.1.1"
mongoengine = "0.24.1"
packaging = "21.3"
passlib = "1.7.4"
pyasn1 = "0.4.8"
pycodestyle = "2.8.0"
pycparser = "2.21"
pydantic = "1.9.1"
pymongo = "4.1.1"
pyparsing = "3.0.9"
python-dotenv = "0.20.0"
python-jose = "3.3.0"
python-multipart = "0.0.5"
redis = "4.3.4"
requests = "2.28.1"
rfc3986 = "1.5.0"
rsa = "4.8"
six = "1.16.0"
sniffio = "1.2.0"
sortedcontainers = "2.4.0"
starlette = "0.19.1"
toml = "0.10.2"
typing-extensions = "4.3.0"
urllib3 = "1.26.9"
uvicorn = "0.18.2"
uvloop = "0.16.0"
websockets = "10.3"
wrapt = "1.14.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
